JSONService is now the default Service as Adaptation has moved to the Configuration
