+++ Example/Example/Modules/Swift/ExampleBar.swift
@@ -3,7 +3,7 @@ import Faro
 class ExampleBar: Bar {
 
     init() {
-        super.init(service: JSONService(configuration: Configuration(baseURL: "http://jsonplaceholder.typicode.com")))
+        super.init(service: Service(configuration: Configuration(baseURL: "http://jsonplaceholder.typicode.com")))
     }
 
     override func serve<M: Mappable>(order: Order, result: (Result<M>) -> ()) {
