+++ Sources/Adaptable.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+public protocol Adaptable: class {
+    func serialize<M: Mappable>(fromDataResult dataResult: Result<M>, jsonResult: (Result <M>) -> ())
+}
